fix substr<-(chr, start, <negative>) -- PR#18228
git-svn-id: https://svn.r-project.org/R/trunk@81136 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/character.c
@@ -765,7 +765,7 @@ SEXP attribute_hidden do_substrgets(SEXP call, SEXP op, SEXP args, SEXP env)
             slen = strlen(ss);
             if (start < 1)
                 start = 1;
-            if (stop > slen)
+            if (stop > (int)slen)
                 stop = (int)slen; /* SBCS optimization */
             if (start > stop)
             {
